implemented Async observer pattern
