+++ QLife/app/src/main/java/engsoc/qlife/utility/Constants.java
@@ -15,6 +15,7 @@ public class Constants {
     public static final String GET_ENG_CLASS = "http://calendar.engineering.queensu.ca/content.php?filter%5B27%5D=";
     public static final String FILTER_ENG_CLASS = "&filter%5B29%5D=&filter%5Bcourse_type%5D=-1&filter%5Bkeyword%5D=&filter%5B32%5D=1&filter%5Bcpage%5D=1&cur_cat_oid=2&expand=&navoid=50&search_database=Filter#acalog_template_course_filter";
     public static final String GET_ROOM_BOOKINGS = "https://queensu.evanced.info/dibsAPI/reservations/";
+    public static final String GET_DATABASE="http://qtap.engsoc.queensu.ca/database/get_database.php";
 
     public static final int TIMEOUT = 5000;
 }
